Find the Maximum Observed Effluent Concentration (MEC)

Find the MEC (no projection) from the observed dataset using methods described in EPA's Technical Support Document for Water Quality-based Toxics Control.

find_mec(qual, result, nd = c("<", "nd", "ND"), simple_output = FALSE)

Arguments

qual
A character vector containing non-detect indicator strings, e.g., "<" or "ND". The strings used to indicate censored status can be edited in the "nd" argument.

result
A numeric vector of concentration measurements.

nd
A list indicating all the censoring flags included in the dataset. Defaults to "<", "nd", and "ND".

simple_output
Logical value. If TRUE, the output columns are concatenated into a single character string (e.g., "<0.2") which can be useful for constructing summary tables.

Value

Dataframe with a qualifier column (character) and a MEC column (numeric).

Examples

# Find observed MEC
cen_result <- c(rep("", 10), rep("<", 10))
result     <- 1:20
find_mec(cen_result, result)

#>   qual result
#> 1          10

cen_result <- rep("<", 20)
find_mec(cen_result, result)

#>   qual result
#> 1    <      1

cen_result <- rep("", 20)
find_mec(cen_result, result)

#>   qual result
#> 1          20

# Demonstrate simplified output
find_mec(cen_result, result, simple_output = TRUE)

#> [1] "20"

# Define a set of custom non-detect flags
cen_result <- c(rep("non-detect", 5), rep("<", 10), rep("mdl", 5))
find_mec(cen_result, result, nd = c("non-detect", "<", "mdl"))

#>   qual result
#> 1    <      1